requirements.txt
numpy>=1.9.0
git+https://github.com/nipy/nipype#egg=nipype
scipy>=0.11
networkx>=1.9
traits>=4.6
python-dateutil>=1.5
rdflib
nipy 
dipy
nibabel>=2.1.0
future>=0.16.0
simplejson>=3.8.0
prov==1.5.0
click>=6.6.0
funcsigs
configparser
pytest>=3.0
mock
pydotplus
pydot>=1.2.3
packaging
med2image
pybids